whatsappBot/requirements.txt

certifi==2020.6.20

contextlib2==0.6.0.post1

dataclasses-json==0.5.2

dephell-pythons==0.1.14

dephell-venvs==0.1.16

docopt==0.6.2
evdev==1.3.0
extras==1.0.0

Flask==1.1.2

html5lib==1.1
httplib2==0.18.1
python-dateutil==2.8.1



prettytable==1.0.1

psutil==5.7.2


pycparser==2.20

pycurl==7.43.0.6

pyparsing==2.4.7



python-xlib==0.28



requests==2.24.0

retrying==1.3.3

tabulate==0.8.7


twilio==6.46.0


urllib3==1.25.10

google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0